Phan quyen theo chuc nang

+++ app/Http/Controllers/DepartmentController.php
@@ -2,17 +2,21 @@
 
 namespace App\Http\Controllers;
 use App\Models\Department;
+
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 class DepartmentController extends Controller
 {
     public function index(){
-
+        
         $departments = Department::paginate(10);
         return view('managers\departments', compact('departments'));
     }
 
     public function insertDepartment(Request $request){
+        if ($request->user()->can('create-department')) {
+    
+   
         $messages=[
                 'name.required'=>'Bạn phải nhập tên phòng bang'
                 ];
@@ -27,14 +31,26 @@ class DepartmentController extends Controller
         $success = 'Thêm phòng ban thành công!';
         $departments = Department::paginate(10);
         return view('managers\departments', compact('departments'))->with('success', $success);
-    }
-    public function delDepartment($id){
+        }
+
+        $alert = 'Bạn không có quyền thêm';
+
+        $departments = Department::paginate(10);
+        return view('managers\departments', compact('departments'))->with('alert', $alert);
+        }
+    public function delDepartment(Request $request,$id){
+        if ($request->user()->can('create-department')){
         $record = Department::where("id", $id)->first();
         
         
         Department::where("id", $id)->delete();
         $department = Department::paginate(10);
         return redirect()->action([DepartmentController::class, 'index']);
+        }
+        $alert = 'Bạn không có quyền xóa';
+
+        $departments = Department::paginate(10);
+        return view('managers\departments', compact('departments'))->with('alert', $alert);
     }
 
     public function edit($id)
